image-upload-with-cloudinary-|-updated-and-renamed-context
