--- client/src/context/AuthContext.js
@@ -1,11 +1,0 @@
-import { createContext } from 'react';
-
-function noop() {}
-
-export const AuthContext = createContext({
-    token: null,
-    userId: null,
-    login: noop,
-    logout: noop,
-    isAuthenticated: false
-})
